backend/requirements.txt

Flask==2.3.3
Flask-Cors==4.0.0
Pillow==10.0.0
ultralytics==8.0.196
numpy==1.24.3
opencv-python==4.8.0.76
python-dotenv==1.0.0
gunicorn==21.2.0
